feat:componet's navBar
