+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "files.exclude": {
+    "src": false
+  }
+}
