+++ index.html
@@ -16,7 +16,7 @@
     />
   </head>
   <body>
-    <nav id="nav"></nav>
+    <header class="header" id="header"></header>
     <script src="/src/js/index.js" type="module"></script>
   </body>
 </html>
