+++ src/consts/Data.js
@@ -1,0 +1,52 @@
+// -------- Nav Data --------
+export const navDataLinks = [
+  {
+    name: "Inicio",
+    href: "#home",
+    icon: "home",
+  },
+  {
+    name: "Sobre",
+    href: "#about",
+    icon: "people_outline",
+  },
+  {
+    name: "Resumo",
+    href: "#resume",
+    icon: "class",
+  },
+  {
+    name: "Projetos",
+    href: "#projects",
+    icon: "work_outline",
+  },
+  {
+    name: "Contato",
+    href: "#contact",
+    icon: "chat",
+  },
+]
+
+// ----- Social redes data -----
+
+export const logoSocial = [
+  {
+    name: "facebook",
+    iconRede: "src/assets/facebook.svg",
+    url: "https://www.facebook.com/lucas.surita.3/",
+  },
+  {
+    name: "github",
+    iconRede: "src/assets/github.svg",
+    url: "https://github.com/Lucas-Silva-Amaral",
+  },
+  {
+    name: "linkedin",
+    iconRede: "src/assets/linkedin.svg",
+    url: "https://www.linkedin.com/in/lucas-amaral-1b174b8b/",
+  },
+]
+
+export const titles = {
+  titleFooter: "Encontre-me",
+}
