--- src/consts/navDataLinks.js
@@ -1,35 +1,0 @@
-export const navDataLinks = [
-  {
-    name: "Inicio",
-    href: "#home",
-    icon: "home",
-  },
-  {
-    name: "Sobre",
-    href: "#about",
-    icon: "people_outline",
-  },
-  {
-    name: "Resumo",
-    href: "#resume",
-    icon: "class",
-  },
-  {
-    name: "Projetos",
-    href: "#projects",
-    icon: "work_outline",
-  },
-  {
-    name: "Contato",
-    href: "#contact",
-    icon: "chat",
-  },
-]
-
-export const profile = {
-  name: "Lucas Silva Amaral",
-  profission: "Desenvolvedor Front-end",
-  img: "src/assets/lucasyo.png",
-  about:
-    " Olá, me chamo Lucas e tenho 26 anos, atualmente sou discente de Análise e Desenvolvimento de Sistemas no Senac Pelotas. Desde pequeno sou apaixonado por tecnologia.Então em 2021 descobri esta área sensacional que é a programação, paixão a primeira vista, desde então estou nesta caminhada em busca de conhecimento",
-}
