+++ src/css/style.css
@@ -1,4 +1,5 @@
 @import url("global/global.css");
 @import url("ultility/colors.css");
 @import url("ultility/fonts.css");
+@import url("../js/moldules/NavBar/navNarHover.css");
 @import url("../screens/NavBar/navBar.css");
