+++ src/css/ultility/colors.css
@@ -6,4 +6,6 @@
   --color-bg: #878e99;
   --color-hover: #ff014f;
   --color-border: #17191c;
+  --shadow-1: 10px 10px 19px #1c1e22, -10px -10px 19px #262a2e;
+  --shadow-2: inset 21px 21px 19px #181a1d, inset -21px -21px 19px #202225;
 }
