+++ src/js/moldules/NavBar/navNarHover.css
@@ -1,0 +1,11 @@
+.ulHover li a,
+.ulHover li i {
+  color: var(--color-bg);
+  transition: all 0.5s ease-in-out;
+}
+
+.linksHoverActive i,
+.linksHoverActive {
+  color: var(--color-hover) !important;
+  transition: all 0.5s ease-in-out;
+}
