+++ src/screens/NavBar/navBar.css
@@ -1,12 +1,21 @@
 .header {
-  max-width: 293px;
+  max-width: 290px;
+  width: 100%;
   height: 100vh;
   display: flex;
   flex-direction: column;
   position: fixed;
   background-color: var(--color-bgHeader);
-  padding: 50px 40px 40px 40px;
+  padding: 50px 40px 40px;
   z-index: 1;
+}
+
+#nav {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: var(--color-bgHeader);
 }
 
 .divLogo {
@@ -18,8 +27,6 @@
 }
 
 .logo {
-  max-width: 400px;
-  max-height: 193px;
   min-width: auto;
   min-height: auto;
   display: block;
@@ -31,6 +38,7 @@
 
 .ulLinks {
   max-width: 193px;
+  width: 100%;
   min-width: auto;
   display: flex;
   flex-direction: column;
@@ -38,10 +46,12 @@
 }
 
 .liLinks {
+  width: 100%;
   display: flex;
   border-bottom: 1px solid var(--color-bgHeader);
   margin-top: 9.5px;
   margin-bottom: 9.5px;
+  justify-content: flex-start;
 }
 
 .liLinks a,
@@ -49,10 +59,10 @@ i {
   color: var(--color-text);
   font-size: 1.06em;
   line-height: 1.5em;
-  padding: 0px 10px;
-  text-decoration: none;
   display: flex;
   align-items: center;
+  gap: 0.5em;
+  transition: all 0.3s ease-in-out;
 }
 
 .divSocial {
@@ -70,14 +80,47 @@ i {
   margin-top: 20px;
 }
 
-.ulSocial a {
-  display: flex;
-  align-items: center;
+.socialLink {
   font-size: 1.06em;
   line-height: 1.5em;
   padding: 10px 15px 15px 15px;
-  border: solid 1px #32363b;
   border-radius: 10px;
   background-color: 145deg, rgb(30, 32, 36);
-  box-shadow: 0 0 10px rgba(0, 0, 0);
+  box-shadow: 0px 0px 10px rgba(0, 0, 0);
+  display: flex;
+  vertical-align: middle;
+  -webkit-transform: perspective(1px) translateZ(0) scale(1);
+  transform: perspective(1px) translateZ(0) scale(1);
+  position: relative;
+  -webkit-transition-duration: 0.5s;
+  transition-duration: 0.5s;
+  -webkit-transition-property: transform;
+  transition-property: transform;
+  box-shadow: var(--shadow-1);
+}
+
+.socialLink:before {
+  pointer-events: none;
+  position: absolute;
+  z-index: -1;
+  content: "";
+  top: 100%;
+  left: 5%;
+  height: 10px;
+  width: 90%;
+  opacity: 0;
+  -webkit-transition-duration: 0.5s;
+  transition-duration: 0.5s;
+  -webkit-transition-property: transform, opacity;
+  transition-property: transform, opacity;
+}
+.socialLink:hover,
+.socialLink:focus,
+.socialLink:active {
+  box-shadow: 5px 10px 10px rgba(0, 0, 0, 0.5);
+  background-color: var(--color-hover);
+  -webkit-transform: translateY(-10px)
+    matrix3d(1, 0, 0, 0, 0.5, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1);
+  transform: translateY(-10px);
+  /* move the element up by 10px */
 }
